+++ src/java/us/temerity/pipeline/math/Color3d.java
@@ -299,7 +299,7 @@ class Color3d
     /* has saturation... */ 
     else {
       double hue  = hsv.x() / 60.0;
-      int region = (int) Math.floor((double) hue);
+      int region = (int) Math.floor(hue);
       double frac = hue - ((double) region);
 
       double p = hsv.z() * (1.0 - hsv.y());
