+++ src/java/us/temerity/pipeline/math/Matrix33d.java
@@ -375,8 +375,8 @@ class Matrix33d
    double theta
   ) 
   {								    
-    double s = (double) Math.sin(theta);
-    double c = (double) Math.cos(theta);
+    double s = Math.sin(theta);
+    double c = Math.cos(theta);
     
     identity();
     pCols[0].setComp(0, c); pCols[0].setComp(1, -s);
